ScrapperJB/requirements.txt
pydantic==2.9.2
tqdm==4.67.1
requests==2.32.3
strawberry-graphql[fastapi]==0.243.0
uvicorn[standard]==0.32.1
fastapi==0.115.6
